Move LatLngConverter to util
