+++ ground/src/main/java/com/google/android/ground/model/submission/CaptureLocationTaskData.kt
@@ -17,8 +17,8 @@ package com.google.android.ground.model.submission
 
 import android.location.Location
 import com.google.android.ground.model.geometry.Coordinates
-import com.google.android.ground.model.geometry.LatLngConverter
 import com.google.android.ground.model.geometry.Point
+import com.google.android.ground.util.LatLngConverter
 import java.math.RoundingMode
 import java.text.DecimalFormat
