+++ ground/src/test/java/com/google/android/ground/util/LatLngConverterTest.kt
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.android.ground.ui.datacollection.tasks.point
+package com.google.android.ground.util
 
 import com.google.android.ground.BaseHiltTest
 import com.google.android.ground.model.geometry.Coordinates
-import com.google.android.ground.model.geometry.LatLngConverter.formatCoordinates
+import com.google.android.ground.util.LatLngConverter.formatCoordinates
 import com.google.common.truth.Truth.assertThat
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Test
